pyproject.toml
[project]
name = "accmsc"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "actipy>=3.5.0",
    "hydra-core>=1.3.2",
    "numpy>=2.2.4",
    "omegaconf>=2.3.0",
    "pandas>=2.2.3",
    "requests>=2.32.3",
    "scipy>=1.15.2",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "torchaudio>=2.6.0",
    "tqdm>=4.67.1",
    "torchmetrics>=1.7.1",
    "matplotlib>=3.10.1",
    "scikit-learn>=1.6.1",
    "hmmlearn>=0.3.3",
    "tensorboard>=2.19.0",
]
